add FeatureSupport
