+++ openpathsampling/storage/snapshot_store.py
@@ -1,21 +1,26 @@
 import numpy as np
 
-from openpathsampling.snapshot import Snapshot, Configuration, Momentum, AbstractSnapshot, ToySnapshot
+from openpathsampling.snapshot import Snapshot, AbstractSnapshot, ToySnapshot
 from openpathsampling.trajectory import Trajectory
 from openpathsampling.netcdfplus import ObjectStore, LoaderProxy
 
-import simtk.unit as u
+import features as ft
+
+from features import ConfigurationStore, MomentumStore
 
 class AbstractSnapshotStore(ObjectStore):
     """
     An ObjectStore for Snapshots in netCDF files.
     """
 
-    def __init__(self, snapshot_class = AbstractSnapshot):
-        super(AbstractSnapshotStore, self).__init__(snapshot_class, json=False)
+    def __init__(self, snapshot_class):
+        super(AbstractSnapshotStore, self).__init__(AbstractSnapshot, json=False)
+        self.snapshot_class = snapshot_class
 
     def to_dict(self):
-        return {}
+        return {
+            'snapshot_class': self.snapshot_class
+        }
 
     def _get(self, idx, from_reversed=False):
         if from_reversed:
@@ -124,7 +129,7 @@ class SnapshotStore(AbstractSnapshotStore):
         if from_reversed:
             obj = self.cache[idx ^ 1]
 
-            return self.content_class(
+            return Snapshot(
                 configuration=obj.configuration,
                 momentum=obj.momentum,
                 is_reversed=not obj.is_reversed,
@@ -135,7 +140,7 @@ class SnapshotStore(AbstractSnapshotStore):
             momentum = self.vars['momentum'][idx]
             momentum_reversed = self.vars['momentum_reversed'][idx]
 
-            return self.content_class(
+            return Snapshot(
                 configuration=configuration,
                 momentum=momentum,
                 is_reversed=momentum_reversed,
@@ -307,173 +312,59 @@ class ToySnapshotStore(AbstractSnapshotStore):
 
         return idx
 
-
-class MomentumStore(ObjectStore):
+class FeatureSnapshotStore(AbstractSnapshotStore):
     """
-    An ObjectStore for Momenta. Allows to store Momentum() instances in a netcdf file.
+    An ObjectStore for Snapshots in netCDF files.
     """
 
-    def __init__(self):
-        super(MomentumStore, self).__init__(Momentum, json=False)
+    def __init__(self, content_class):
+        super(FeatureSnapshotStore, self).__init__(content_class)
 
-    def _save(self, momentum, idx):
-        self.vars['velocities'][idx, :, :] = momentum.velocities
+    @property
+    def features(self):
+        return self.content_class.__features__
 
-        if momentum.kinetic_energy is not None:
-            self.vars['kinetic_energy'][idx] = momentum.kinetic_energy
+    def to_dict(self):
+        return {
+            'snapshot_class': self.content_class
+        }
 
-    def _load(self, idx):
-        velocities = self.vars['velocities'][idx]
-        kinetic_energy = self.vars['kinetic_energy'][idx]
+    def _put(self, idx, snapshot):
+        for feature in self.features:
+            self.vars[feature][idx] = getattr(snapshot, feature)
+            self.write(feature, idx ^ 1, snapshot)
 
-        momentum = Momentum(velocities=velocities, kinetic_energy=kinetic_energy)
-
-        return momentum
-
-    def velocities_as_numpy(self, frame_indices=None, atom_indices=None):
-        """
-        Return a block of stored velocities in the database as a numpy array.
-
-        Parameters
-        ----------
-        frame_indices : list of int or None
-            the indices of Momentum objects to be retrieved from the database.
-            If `None` is specified then all indices are returned!
-        atom_indices : list of int of None
-            if not None only the specified atom_indices are returned. Might
-            speed up reading a lot.
-        """
-
-        if frame_indices is None:
-            frame_indices = slice(None)
-
-        if atom_indices is None:
-            atom_indices = slice(None)
-
-        return self.variables['velocities'][frame_indices, atom_indices, :].astype(np.float32).copy()
-
-    def velocities_as_array(self, frame_indices=None, atom_indices=None):
-        """
-        Returns a numpy array consisting of all velocities at the given indices
-
-        Parameters
-        ----------
-        frame_indices : list of int
-            momenta indices to be loaded
-        atom_indices : list of int
-            selects only the atoms to be returned. If None (Default) all atoms
-            will be selected
+        self.vars['momentum_reversed'][idx] = snapshot.is_reversed
+        self.vars['momentum_reversed'][idx ^ 1] = not snapshot.is_reversed
 
 
-        Returns
-        -------
-        numpy.ndarray, shape = (l,n)
-            returns an array with `l` the number of frames and `n` the number
-            of atoms
-        """
+    def _get(self, idx, from_reversed=False):
+        if from_reversed:
+            obj = self.cache[idx ^ 1]
 
-        return self.velocities_as_numpy(frame_indices, atom_indices)
+            snapshot = self.snapshot_class.__new__(self.content_class)
+
+            for feature in self.features:
+                setattr(snapshot, feature, getattr(obj, feature))
+
+            snapshot.is_reversed = not obj.is_reversed
+        else:
+            snapshot = self.snapshot_class.__new__(self.content_class)
+
+            for feature in self.features:
+                setattr(snapshot, feature, self.vars[feature][idx])
+
+                coordinates = self.vars['coordinates'][idx]
+                velocities = self.vars['velocities'][idx]
+
+            snapshot.is_reversed = self.vars['momentum_reversed'][idx]
+
+        snapshot.topology = self.storage.topology
+        snapshot._reversed = LoaderProxy(self, idx ^ 1)
+
 
     def _init(self):
-        """
-        Initializes the associated storage to index momentums in it
-        """
+        super(FeatureSnapshotStore, self)._init()
 
-        super(MomentumStore, self)._init()
-
-        n_atoms = self.storage.n_atoms
-        n_spatial = self.storage.n_spatial
-
-        self.init_variable('velocities', 'numpy.float32',
-                           dimensions=('atom', 'spatial'),
-                           description="the velocity of atom 'atom' in dimension " +
-                                       "'coordinate' of momentum 'momentum'.",
-                           chunksizes=(1, n_atoms, n_spatial),
-                           simtk_unit=u.nanometers / u.picoseconds
-                           )
-
-        self.init_variable('kinetic_energy', 'float',
-                           chunksizes=(1,),
-                           simtk_unit=u.kilocalories_per_mole
-                           )
-
-
-class ConfigurationStore(ObjectStore):
-    def __init__(self):
-        super(ConfigurationStore, self).__init__(Configuration, json=False)
-
-    def _save(self, configuration, idx):
-        # Store configuration.
-        self.vars['coordinates'][idx] = configuration.coordinates
-
-        if configuration.potential_energy is not None:
-            self.vars['potential_energy'][idx] = configuration.potential_energy
-
-        if configuration.box_vectors is not None:
-            self.vars['box_vectors'][idx] = configuration.box_vectors
-
-    def get(self, indices):
-        return [self.load(idx) for idx in indices]
-
-    def _load(self, idx):
-        coordinates = self.vars["coordinates"][idx]
-        box_vectors = self.vars["box_vectors"][idx]
-        potential_energy = self.vars["potential_energy"][idx]
-
-        configuration = Configuration(coordinates=coordinates, box_vectors=box_vectors,
-                                      potential_energy=potential_energy)
-        configuration.topology = self.storage.topology
-
-        return configuration
-
-    def coordinates_as_numpy(self, frame_indices=None, atom_indices=None):
-        """
-        Return the atom coordinates in the storage for given frame indices
-        and atoms
-
-        Parameters
-        ----------
-        frame_indices : list of int or None
-            the frame indices to be included. If None all frames are returned
-        atom_indices : list of int or None
-            the atom indices to be included. If None all atoms are returned
-
-        Returns
-        -------
-        numpy.array, shape=(n_frames, n_atoms)
-            the array of atom coordinates in a float32 numpy array
-
-        """
-        if frame_indices is None:
-            frame_indices = slice(None)
-
-        if atom_indices is None:
-            atom_indices = slice(None)
-
-        return self.storage.variables[self.prefix + '_coordinates'][frame_indices, atom_indices, :].astype(
-            np.float32).copy()
-
-    def _init(self):
-        super(ConfigurationStore, self)._init()
-        n_atoms = self.storage.n_atoms
-        n_spatial = self.storage.n_spatial
-
-        self.init_variable('coordinates', 'numpy.float32',
-                           dimensions=('atom', 'spatial'),
-                           description="coordinate of atom '{ix[1]}' in dimension " +
-                                       "'{ix[2]}' of configuration '{ix[0]}'.",
-                           chunksizes=(1, n_atoms, n_spatial),
-                           simtk_unit=u.nanometers
-                           )
-
-        self.init_variable('box_vectors', 'numpy.float32',
-                           dimensions=('spatial', 'spatial'),
-                           chunksizes=(1, n_spatial, n_spatial),
-                           simtk_unit=u.nanometers
-                           )
-
-        self.init_variable('potential_energy', 'float',
-                           chunksizes=(1,),
-                           simtk_unit=u.kilocalories_per_mole
-                           )
-
+        for feature in self.features:
+            getattr(ft, feature)._init(self)
